bug: Stops input being overriden on item search
Prior to this change, I'd setup an after reflex hook that would trigger
the focus to be updated on the element with the correct attribute.
However this causes reflex to fire everytime I typed, which also
triggers a reflex action, causing the input to constantly being cleared
after each letter

This change, scopes the reflex hook to only trigger inside the list item
stimulus controller and only for a particular reflex.